Updates to find_postk function
Thank you for posting this code.  While I was applying it to one of my projects, I noted that the find_postk function was a bottleneck to a simulation study I was conducting.  After reviewing the code, I determined that several of the functions within find_postk were using loops (and, in one case, a nested loop) which tends to be inefficient in R.  Whenever possible, I converted the calculations into matrix operations.  For example, I replaced the calculation of loglik with the line
  loglik = ((rM*log(phiM*binvec+ploM*(1-binvec)))%*%rep(1,kk))  +(((nM-rM)*log((1-phiM)*binvec+(1-ploM)*(1-binvec)))%*%rep(1,kk))
which avoids the loop through of the 2^kk individual states.  Based on my estimates, this increases the efficiency of calculation by 80%.